catagory page generation

+++ class/delete_user.php
@@ -1,7 +1,8 @@
 <?php
 include_once('db_functions.php');
 
-if(isset($_GET['username'])){
+//wont delete an admin. Admins can only be deleted using the web host
+if(isset($_GET['username']) && $_GET['RoleID'] != 1){
 	$username = $_GET['username'];
 	delete_row('users','username',$username);
 	header("Location: admin_control.php");
